requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
python-docx==1.1.0
python-pptx==0.6.23
google-genai
pdf2image>=1.16.0